Reorganized `windows.py` to `quran` folder
