+++ bless/main.py
@@ -1,6 +1,6 @@
 from PySide6 import QtWidgets
 
-from windows import BlessWindow
+from quran.windows import BlessWindow
 
 import sys
