+++ bless/quran/windows.py
